+++ src/services/permission-service.ts
@@ -1,19 +1,21 @@
-import { Service } from "./service";
-import { settings } from "../settings";
+import { Service, LoginInfo } from "./service";
 import { errors } from "../errors";
 import { User, Resource, ResourceType, Role } from "../models";
+import { events } from "../events";
 
 export class PermissionService extends Service {
+
+    static baseUrl: string
 
     constructor() {
         super()
     }
 
     protected url(path: string) {
-        if (!settings.permissionServiceUrl)
+        if (!PermissionService.baseUrl)
             throw errors.serviceUrlCanntNull('permissionService')
 
-        return `${settings.permissionServiceUrl}/${path}`
+        return `${PermissionService.baseUrl}/${path}`
     }
 
     //=============================================================
@@ -21,6 +23,8 @@ export class PermissionService extends Service {
 
     /** 添加资源 */
     async addResource(item: Partial<Resource>) {
+        if (!item) throw errors.argumentNull('item')
+
         let url = this.url('resource/add')
         let result = await this.postByJson<{ id: string }>(url, { item })
         Object.assign(item, result)
@@ -29,6 +33,8 @@ export class PermissionService extends Service {
 
     /** 更新资源 */
     async updateResource(item: Partial<Resource>) {
+        if (!item) throw errors.argumentNull('item')
+
         let url = this.url('resource/update')
         let result = await this.postByJson(url, { item })
         Object.assign(item, result)
@@ -46,6 +52,8 @@ export class PermissionService extends Service {
 
     /** 获取资源列表 */
     async getResourceList(args: DataSourceSelectArguments): Promise<DataSourceSelectResult<Resource>> {
+        if (!args) throw errors.argumentNull('args')
+
         let url = this.url('resource/list')
         if (!args.sortExpression)
             args.sortExpression = 'sort_number asc'
@@ -67,6 +75,8 @@ export class PermissionService extends Service {
      * @param id 要删除的资源编号
      */
     async deleteResource(id: string) {
+        if (!id) throw errors.argumentNull('id')
+
         let url = this.url('resource/remove')
         return this.postByJson(url, { id })
     }
@@ -76,6 +86,8 @@ export class PermissionService extends Service {
      * @param id 资源编号
      */
     async getResourceChildCommands(id: string) {
+        if (!id) throw errors.argumentNull('id')
+
         let buttonType: ResourceType = 'button'
         let filter = `parent_id = '${id}' and type = '${buttonType}'`
         let url = `resource/list`
@@ -100,6 +112,8 @@ export class PermissionService extends Service {
      * @param id 要获取的角色编号
      */
     getRole(id: string): Promise<Role | null> {
+        if (!id) throw errors.argumentNull('id')
+
         let url = this.url('role/get')
         return this.getByJson(url, { id })
     }
@@ -130,6 +144,9 @@ export class PermissionService extends Service {
 
     /** 设置用户角色 */
     setUserRoles(userId: string, roleIds: string[]) {
+        if (!userId) throw errors.argumentNull('userId')
+        if (!roleIds) throw errors.argumentNull('roleIds')
+
         let url = this.url('user/setRoles')
         return this.postByJson(url, { userId, roleIds })
     }
@@ -138,7 +155,7 @@ export class PermissionService extends Service {
     // 用户相关
 
     /** 获取用户列表 */
-    async getUserList(args: DataSourceSelectArguments): Promise<DataSourceSelectResult<User>> {
+    async getUserList(args?: DataSourceSelectArguments): Promise<DataSourceSelectResult<User>> {
         let url = this.url('user/list')
         let result = await this.getByJson<DataSourceSelectResult<User>>(url, { args })
         if (result == null)
@@ -149,6 +166,8 @@ export class PermissionService extends Service {
 
     /** 通过手机获取用户 */
     async getUserByMobile(mobile: string) {
+        if (!mobile) throw errors.argumentNull('mobile')
+
         let args: DataSourceSelectArguments = {}
         args.filter = `mobile = '${mobile}'`
         let r = await this.getUserList(args)
@@ -160,6 +179,8 @@ export class PermissionService extends Service {
      * @param userId 要移除的用户编号
      */
     async removeUser(userId: string) {
+        if (!userId) throw errors.argumentNull('userId')
+
         let url = this.url('application/removeUser')
         return this.deleteByJson(url, { userId })
     }
@@ -175,11 +196,169 @@ export class PermissionService extends Service {
         if (result == null)
             throw errors.unexpectedNullResult()
 
-        // for (let i = 0; i < result.dataItems.length; i++) {
-        //     result.dataItems[i].sort_number = (args.startRowIndex || 0) + i + 1
-        // }
-
         return result
+    }
+
+    /** 
+     * 发送注册操作验证码
+     * @param mobile 接收验证码的手机号
+     */
+    sendRegisterVerifyCode(mobile: string) {
+        let url = this.url('sms/sendVerifyCode')
+        return this.postByJson<{ smsId: string }>(url, { mobile, type: 'register' })
+    }
+
+
+    /**
+     * 校验验证码
+     * @param smsId 验证码信息的 ID 号
+     * @param verifyCode 验证码
+     */
+    async checkVerifyCode(smsId: string, verifyCode: string) {
+        if (!smsId) throw errors.argumentNull('smsId')
+        if (!verifyCode) throw errors.argumentNull('verifycode')
+
+        let url = this.url('sms/checkVerifyCode')
+        let r = await this.postByJson<boolean>(url, { smsId, verifyCode })
+        return r
+    }
+
+    /**
+     * 发送重置密码操作验证码
+     * @param mobile 接收验证码的手机号
+     */
+    sendResetVerifyCode(mobile: string) {
+        if (!mobile) throw errors.argumentNull('mobile')
+
+        let url = this.url('sms/sendVerifyCode')
+        return this.postByJson<{ smsId: string }>(url, { mobile, type: 'resetPassword' })
+    }
+
+    /**
+     * 重置密码
+     * @param mobile 手机号
+     * @param password 新密码
+     * @param smsId 短信编号
+     * @param verifyCode 验证码
+     */
+    resetPassword(mobile: string, password: string, smsId: string, verifyCode: string) {
+        if (!mobile) throw errors.argumentNull('mobile')
+        if (!password) throw errors.argumentNull('password')
+        if (!smsId) throw errors.argumentNull('smsId')
+        if (!verifyCode) throw errors.argumentNull('verifyCode')
+
+        let url = this.url('user/resetPassword')
+        return this.postByJson(url, { mobile, password, smsId, verifyCode })
+    }
+
+    /**
+     * 重置手机号码
+     * @param mobile 需要重置的新手机号
+     * @param smsId 短信编号
+     * @param verifyCode 验证码
+     */
+    resetMobile(mobile: string, smsId: string, verifyCode: string) {
+        if (!mobile) throw errors.argumentNull('mobile')
+        if (!smsId) throw errors.argumentNull('smsId')
+        if (!verifyCode) throw errors.argumentNull('verifyCode')
+
+        let url = this.url('user/resetMobile')
+        return this.postByJson(url, { mobile, smsId, verifyCode })
+    }
+
+    /**
+     * 退出登录
+     */
+    logout() {
+        if (Service.loginInfo.value == null)
+            return
+
+        //TODO: 将服务端 token 设置为失效
+
+        events.logout.fire(this, Service.loginInfo.value)
+        Service.setStorageLoginInfo(null)
+        Service.loginInfo.value = null
+    }
+
+    /**
+     * 登录
+     * @param username 用户名
+     * @param password 密码
+     */
+    async login(username: string, password: string) {
+        if (!username) throw errors.argumentNull('username')
+        if (!password) throw errors.argumentNull('password')
+
+        let url = this.url('user/login')
+        let r = await this.postByJson<LoginInfo | null>(url, { username, password })
+        if (r == null)
+            throw errors.unexpectedNullResult()
+
+        Service.loginInfo.value = r
+        Service.setStorageLoginInfo(r)
+        events.login.fire(this, r)
+        return r
+    }
+
+    /**
+     * 注册
+     * @param mobile 手机号
+     * @param password 密码
+     * @param smsId 短信编号
+     * @param verifyCode 验证码
+     */
+    async register(mobile: string, password: string, smsId: string, verifyCode: string, data?: { [key: string]: any }) {
+        if (!mobile) throw errors.argumentNull('mobile')
+        if (!password) throw errors.argumentNull('password')
+        if (!smsId) throw errors.argumentNull('smsId')
+        if (!verifyCode) throw errors.argumentNull('verifyCode')
+
+        let url = this.url('user/register')
+        let r = await this.postByJson<LoginInfo>(url, { mobile, password, smsId, verifyCode, data })
+        if (r == null)
+            throw errors.unexpectedNullResult()
+
+        Service.setStorageLoginInfo(r)
+        events.register.fire(this, r)
+
+        return r;
+    }
+
+    /**
+     * 获取用户个人信息
+     */
+    async me() {
+        let url = this.url('user/me')
+        let user = await this.getByJson<User>(url)
+        return user
+    }
+
+    /**
+     * 获取用户
+     * @param userId 用户编号
+     */
+    async getUser(userId: string) {
+        let url = this.url('user/item')
+        let user = await this.getByJson<User | null>(url, { userId })
+        return user
+    }
+
+    /**
+     * 更新用户信息
+     * @param user 用户
+     */
+    update(user: User) {
+        let url = this.url('user/update')
+        return this.postByJson(url, { user })
+    }
+
+    /**
+     * 获取当前登录用户的角色
+     */
+    async myRoles() {
+        let url = this.url('user/getRoles')
+        let roles = await this.getByJson<Role[]>(url)
+        return roles
     }
 }
